# multiple/multiple_sample/api/v1/gen/go/gateway/greeter_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.8.4
// - protoc             (unknown)
// source: gateway/greeter.proto

package gateway

import (
	helloworld "basic-layout/multiple/multiple_sample/api/v1/gen/go/helloworld"
	secondworld "basic-layout/multiple/multiple_sample/api/v1/gen/go/secondworld"
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationGatewayAPISayHello = "/api.v1.services.gateway.GatewayAPI/SayHello"
const OperationGatewayAPISaySecond = "/api.v1.services.gateway.GatewayAPI/SaySecond"

type GatewayAPIHTTPServer interface {
	// SayHello Forwards to helloworld.HelloGreeterAPI.SayHello
	SayHello(context.Context, *helloworld.SayHelloRequest) (*helloworld.SayHelloResponse, error)
	// SaySecond Forwards to secondworld.SecondGreeterAPI.SayHello
	SaySecond(context.Context, *secondworld.SayHelloRequest) (*secondworld.SayHelloResponse, error)
}

func RegisterGatewayAPIHTTPServer(s *http.Server, srv GatewayAPIHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/gateway/hello/{id}", _GatewayAPI_SayHello0_HTTP_Handler(srv))
	r.GET("/api/v1/gateway/second/{id}", _GatewayAPI_SaySecond0_HTTP_Handler(srv))
}

func _GatewayAPI_SayHello0_HTTP_Handler(srv GatewayAPIHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in helloworld.SayHelloRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGatewayAPISayHello)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SayHello(ctx, req.(*helloworld.SayHelloRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*helloworld.SayHelloResponse)
		return ctx.Result(200, reply)
	}
}

func _GatewayAPI_SaySecond0_HTTP_Handler(srv GatewayAPIHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in secondworld.SayHelloRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGatewayAPISaySecond)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SaySecond(ctx, req.(*secondworld.SayHelloRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*secondworld.SayHelloResponse)
		return ctx.Result(200, reply)
	}
}

type GatewayAPIHTTPClient interface {
	SayHello(ctx context.Context, req *helloworld.SayHelloRequest, opts ...http.CallOption) (rsp *helloworld.SayHelloResponse, err error)
	SaySecond(ctx context.Context, req *secondworld.SayHelloRequest, opts ...http.CallOption) (rsp *secondworld.SayHelloResponse, err error)
}

type GatewayAPIHTTPClientImpl struct {
	cc *http.Client
}

func NewGatewayAPIHTTPClient(client *http.Client) GatewayAPIHTTPClient {
	return &GatewayAPIHTTPClientImpl{client}
}

func (c *GatewayAPIHTTPClientImpl) SayHello(ctx context.Context, in *helloworld.SayHelloRequest, opts ...http.CallOption) (*helloworld.SayHelloResponse, error) {
	var out helloworld.SayHelloResponse
	pattern := "/api/v1/gateway/hello/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGatewayAPISayHello))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *GatewayAPIHTTPClientImpl) SaySecond(ctx context.Context, in *secondworld.SayHelloRequest, opts ...http.CallOption) (*secondworld.SayHelloResponse, error) {
	var out secondworld.SayHelloResponse
	pattern := "/api/v1/gateway/second/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGatewayAPISaySecond))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}