# multiple/multiple_sample/resources/docs/openapi/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Basic Layout Example API
    description: An example project demonstrating a multi-module gRPC backend with a gateway, based on the OrigAdmin framework.
    contact:
        name: OrigAdmin Community
        url: https://github.com/basic-layout/multiple/multiple_sample
        email: waitforadding@gmail.com
    license:
        name: MIT
        url: https://github.com/basic-layout/multiple/multiple_sample/blob/main/LICENSE
    version: v1.0.0
servers:
    - url: http://localhost:8000
    - url: https://localhost:8000
paths:
    /api/v1/gateway/hello/{id}:
        get:
            tags:
                - GatewayAPI
            description: Forwards to helloworld.HelloGreeterAPI.SayHello
            operationId: GatewayAPI_SayHello
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
                - name: data.id
                  in: query
                  description: |-
                    The original proto had an entgo option here, which is likely not needed for this example.
                     option (ent.schema).gen = true;
                  schema:
                    type: string
                - name: data.name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.helloworld.SayHelloResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/gateway/second/{id}:
        get:
            tags:
                - GatewayAPI
            description: Forwards to secondworld.SecondGreeterAPI.SayHello
            operationId: GatewayAPI_SaySecond
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
                - name: data.id
                  in: query
                  description: |-
                    The original proto had an entgo option here, which is likely not needed for this example.
                     option (ent.schema).gen = true;
                  schema:
                    type: string
                - name: data.name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.secondworld.SayHelloResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/helloworld/greeter:
        get:
            tags:
                - HelloGreeterAPI
            description: ListGreeter lists Greeters
            operationId: HelloGreeterAPI_ListGreeter
            parameters:
                - name: current
                  in: query
                  description: Current page
                  schema:
                    type: integer
                    format: int32
                - name: page_size
                  in: query
                  description: The number of lines per page
                  schema:
                    type: integer
                    format: int32
                - name: no_paging
                  in: query
                  description: Whether not paging
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.helloworld.ListGreeterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        post:
            tags:
                - HelloGreeterAPI
            description: CreateGreeter creates a new Greeter
            operationId: HelloGreeterAPI_CreateGreeter
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.helloworld.GreeterData'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.helloworld.CreateGreeterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/helloworld/greeter/{id}:
        get:
            tags:
                - HelloGreeterAPI
            description: GetGreeter gets a Greeter
            operationId: HelloGreeterAPI_GetGreeter
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.helloworld.GetGreeterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        put:
            tags:
                - HelloGreeterAPI
            description: UpdateGreeter updates a Greeter
            operationId: HelloGreeterAPI_UpdateGreeter
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.helloworld.GreeterData'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.helloworld.UpdateGreeterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        delete:
            tags:
                - HelloGreeterAPI
            description: DeleteGreeter deletes a Greeter
            operationId: HelloGreeterAPI_DeleteGreeter
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.helloworld.DeleteGreeterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/helloworld/greeter/{id}/hello:
        get:
            tags:
                - HelloGreeterAPI
            description: SayHello is a get method
            operationId: HelloGreeterAPI_SayHello
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
                - name: data.id
                  in: query
                  description: |-
                    The original proto had an entgo option here, which is likely not needed for this example.
                     option (ent.schema).gen = true;
                  schema:
                    type: string
                - name: data.name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.helloworld.SayHelloResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        post:
            tags:
                - HelloGreeterAPI
            description: PostHello is a post method
            operationId: HelloGreeterAPI_PostHello
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.helloworld.GreeterData'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.helloworld.PostHelloResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/secondworld/greeter:
        get:
            tags:
                - SecondGreeterAPI
            description: ListGreeter lists Greeters
            operationId: SecondGreeterAPI_ListGreeter
            parameters:
                - name: current
                  in: query
                  description: Current page
                  schema:
                    type: integer
                    format: int32
                - name: page_size
                  in: query
                  description: The number of lines per page
                  schema:
                    type: integer
                    format: int32
                - name: no_paging
                  in: query
                  description: Whether not paging
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.secondworld.ListGreeterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        post:
            tags:
                - SecondGreeterAPI
            description: CreateGreeter creates a new Greeter
            operationId: SecondGreeterAPI_CreateGreeter
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.secondworld.GreeterData'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.secondworld.CreateGreeterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/secondworld/greeter/{id}:
        get:
            tags:
                - SecondGreeterAPI
            description: GetGreeter gets a Greeter
            operationId: SecondGreeterAPI_GetGreeter
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.secondworld.GetGreeterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        put:
            tags:
                - SecondGreeterAPI
            description: UpdateGreeter updates a Greeter
            operationId: SecondGreeterAPI_UpdateGreeter
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.secondworld.GreeterData'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.secondworld.UpdateGreeterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        delete:
            tags:
                - SecondGreeterAPI
            description: DeleteGreeter deletes a Greeter
            operationId: SecondGreeterAPI_DeleteGreeter
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.secondworld.DeleteGreeterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/secondworld/greeter/{id}/hello:
        get:
            tags:
                - SecondGreeterAPI
            description: SayHello is a get method
            operationId: SecondGreeterAPI_SayHello
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
                - name: data.id
                  in: query
                  description: |-
                    The original proto had an entgo option here, which is likely not needed for this example.
                     option (ent.schema).gen = true;
                  schema:
                    type: string
                - name: data.name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.secondworld.SayHelloResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        post:
            tags:
                - SecondGreeterAPI
            description: PostHello is a post method
            operationId: SecondGreeterAPI_PostHello
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.secondworld.GreeterData'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.secondworld.PostHelloResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
components:
    schemas:
        api.v1.services.helloworld.CreateGreeterResponse:
            type: object
            properties: {}
        api.v1.services.helloworld.DeleteGreeterResponse:
            type: object
            properties: {}
        api.v1.services.helloworld.GetGreeterResponse:
            type: object
            properties: {}
        api.v1.services.helloworld.GreeterData:
            type: object
            properties:
                id:
                    type: string
                    description: |-
                        The original proto had an entgo option here, which is likely not needed for this example.
                         option (ent.schema).gen = true;
                name:
                    type: string
        api.v1.services.helloworld.ListGreeterResponse:
            type: object
            properties:
                success:
                    type: boolean
                total:
                    type: integer
                    format: int32
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.helloworld.GreeterData'
                error:
                    type: integer
                    description: |-
                        Updated to use the new error reason enum from the runtime framework.
                         Note: This is a breaking API change from the original Error message.
                    format: enum
                extra:
                    $ref: '#/components/schemas/google.protobuf.Any'
        api.v1.services.helloworld.PostHelloResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.v1.services.helloworld.GreeterData'
            description: The response message containing the greetings
        api.v1.services.helloworld.SayHelloResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.v1.services.helloworld.GreeterData'
            description: The response message containing the greetings
        api.v1.services.helloworld.UpdateGreeterResponse:
            type: object
            properties:
                success:
                    type: boolean
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.helloworld.GreeterData'
        api.v1.services.secondworld.CreateGreeterResponse:
            type: object
            properties: {}
        api.v1.services.secondworld.DeleteGreeterResponse:
            type: object
            properties: {}
        api.v1.services.secondworld.GetGreeterResponse:
            type: object
            properties: {}
        api.v1.services.secondworld.GreeterData:
            type: object
            properties:
                id:
                    type: string
                    description: |-
                        The original proto had an entgo option here, which is likely not needed for this example.
                         option (ent.schema).gen = true;
                name:
                    type: string
        api.v1.services.secondworld.ListGreeterResponse:
            type: object
            properties:
                success:
                    type: boolean
                total:
                    type: integer
                    format: int32
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.secondworld.GreeterData'
                error:
                    type: integer
                    description: Updated to use the new error reason enum from the runtime framework.
                    format: enum
                extra:
                    $ref: '#/components/schemas/google.protobuf.Any'
        api.v1.services.secondworld.PostHelloResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.v1.services.secondworld.GreeterData'
            description: The response message containing the greetings
        api.v1.services.secondworld.SayHelloResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.v1.services.secondworld.GreeterData'
            description: The response message containing the greetings
        api.v1.services.secondworld.UpdateGreeterResponse:
            type: object
            properties:
                success:
                    type: boolean
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.secondworld.GreeterData'
        google.protobuf.Any:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        google.rpc.Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/google.protobuf.Any'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
    securitySchemes:
        Bearer:
            type: apiKey
            name: Authorization
            in: header
tags:
    - name: GatewayAPI
      description: GatewayAPI acts as a facade for helloworld and secondworld services.
    - name: HelloGreeterAPI
      description: The greeting service definition.
    - name: SecondGreeterAPI
      description: The greeting service definition.